webapp/betterbots/requirements.txt
absl-py==1.2.0
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.2.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
dataclasses==0.6
deepface==0.0.75
dj-database-url==1.0.0
Django==4.0.6
django-heroku==0.3.1
elastic-transport==8.0.1
elasticsearch==8.0.0
fake-useragent==0.1.11
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
flatbuffers==1.12
frozenlist==1.3.0
gast==0.4.0
gdown==4.5.1
geographiclib==1.52
geopy==2.2.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletransx==2.4.2
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mtcnn==0.1.1
multidict==6.0.2
numpy==1.23.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
Pillow==9.2.0
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.3
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
schedule==1.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tqdm==4.64.0
twint==2.1.21
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.8
Werkzeug==2.2.2
whitenoise==6.2.0
wincertstore==0.2
windows-curses==2.3.0
wrapt==1.14.1
yarl==1.7.2
